move save and load functions to utils (#736)
* move save load to utils

* fix saveload test
