+++ neuralprophet/__init__.py
@@ -27,4 +27,5 @@ from ._version import __version__
 
 from .forecaster import NeuralProphet
 from .utils import set_random_seed, set_log_level
+from .utils import save, load
 from .df_utils import split_df
