+++ neuralprophet/forecaster.py
@@ -2900,41 +2900,3 @@ class NeuralProphet:
                         yhat_df = pd.Series(yhat, name=comp).set_axis(df_forecast.index)
                         df_forecast = pd.concat([df_forecast, yhat_df], axis=1, ignore_index=False)
         return df_forecast
-
-
-def save(model, path):
-    """save a fitted np model to a disk file.
-
-    Parameters
-    ----------
-        model : np.forecaster.NeuralProphet
-            input model that is fitted
-        path : str
-            path and filename to be saved. filename could be any but suggested to have extension .np.
-    Examples
-    --------
-    After you fitted a model, you may save the model to save_test_model.np
-        >>> forecaster.save(model, "test_save_model.np")
-    """
-    torch.save(model, path)
-
-
-def load(path):
-    """retrieve a fitted model from a .np file that was saved by save.
-
-    Parameters
-    ----------
-        path : str
-            path and filename to be saved. filename could be any but suggested to have extension .np.
-
-    Returns
-    -------
-        np.forecaster.NeuralProphet
-            a fitted model
-
-    Examples
-    --------
-    Saved model could be loaded from disk file test_save_model.np
-        >>> model = forecaster.load("test_save_model.np")
-    """
-    return torch.load(path)
